add subscription filter

+++ src/main/java/io/java/remove/IMDB_MDS/model/Subscription.java
@@ -1,4 +1,4 @@
 package io.java.remove.IMDB_MDS.model;
 
-public record Subscription(SubscriptionType subscriptionType,Long limit) {
+public record Subscription(SubscriptionType subscriptionType,Integer limit) {
 }
